Missing variable in filenames.csv
The variable "year" is missing in the default filenames.csv file.
This results in a default year of 2020 being applied to any
simulation that uses the COLUMN NAMES in a time series option run.
I have added the variable "year" which lines up with the variable
in the cols list in SCOPE.m, so it can be read in appropriately.